fix(copy): minor alteration
If you're reading this, it probably means you came from a link in a
cover letter, but you're also a technical person who knows what a commit
it. I've made this alteration because the previous wording highlighted
my inability, and I think a recruiter will see that negatively. I hope
that other engineers will see this in a different light, that I am
learning incredibly quickly. Hence, I'm not going to clobber my git
history to expunge this change.